feat: use campaign tracker in the plugin
