+++ packages/plugin-page-view-tracking/src/page-view-tracking.ts
@@ -9,7 +9,7 @@ import {
   PluginType,
   Storage,
 } from '@amplitude/analytics-types';
-import { AdvancedCampaignTracker } from './advanced-campaign-tracker';
+import { PluginCampaignTracker } from './plugin-campaign-tracker';
 import { PageTrackingBrowserOptions } from './typings/page-view-tracking';
 
 export class PageViewTrackingPlugin implements BeforePlugin {
@@ -37,12 +37,8 @@ export class PageViewTrackingPlugin implements BeforePlugin {
     // Share cookie storage with user session storage
     const storage = config.cookieStorage as unknown as Storage<Campaign>;
 
-    const campaignTracker = new AdvancedCampaignTracker(config.apiKey, {
+    const campaignTracker = new PluginCampaignTracker(config.apiKey, storage, {
       ...attributionConfig,
-      trackPageViews: false,
-      track: /* istanbul ignore next */ () => Promise.resolve(),
-      onNewCampaign: /* istanbul ignore next */ () => () => undefined,
-      storage,
     });
 
     const filter = pageTrackingConfig.filter;
