--- packages/plugin-page-view-tracking/src/typings/advanced-campaign-tracker.ts
@@ -1,7 +1,0 @@
-import { Campaign, CampaignTrackerOptions } from '@amplitude/analytics-types';
-
-export interface CampaignTracker extends CampaignTrackerOptions {
-  onPageChange: (
-    callback: (state: { isNewCampaign: boolean; currentCampaign: Campaign }) => Promise<unknown>,
-  ) => Promise<void>;
-}
