+++ packages/plugin-page-view-tracking/test/plugin-campaign-tracker.test.ts
@@ -1,20 +1,17 @@
 import { MemoryStorage } from '@amplitude/analytics-core';
-import { AdvancedCampaignTracker } from '../src/advanced-campaign-tracker';
+import { PluginCampaignTracker } from '../src/plugin-campaign-tracker';
 import { Campaign } from '@amplitude/analytics-types';
 
-describe('AdvancedCampaignTracker', () => {
+describe('PluginCampaignTracker', () => {
   const API_KEY = 'API_KEY';
 
   describe('onPageChange', () => {
     test('should trigger callback', async () => {
       const config = {
         trackPageViews: false,
-        storage: new MemoryStorage<Campaign>(),
-        track: jest.fn(),
-        onNewCampaign: jest.fn(),
       };
-
-      const tracker = new AdvancedCampaignTracker(API_KEY, config);
+      const storage = new MemoryStorage<Campaign>();
+      const tracker = new PluginCampaignTracker(API_KEY, storage, config);
       await tracker.onPageChange(async ({ isNewCampaign }) => {
         expect(isNewCampaign).toBe(false);
       });
