--- packages/plugin-web-attribution/src/advanced-campaign-tracker.ts
@@ -1,43 +1,0 @@
-import { Campaign, CampaignTrackerOptions } from '@amplitude/analytics-types';
-import { CampaignTracker } from '@amplitude/analytics-client-common';
-import { BASE_CAMPAIGN } from './constant';
-
-export class AdvancedCampaignTracker extends CampaignTracker {
-  private _currentCampaign: Campaign = {
-    ...BASE_CAMPAIGN,
-  };
-  private _isNewCampaign = true;
-  ready: Promise<unknown>;
-
-  constructor(apiKey: string, options: CampaignTrackerOptions) {
-    super(apiKey, options);
-
-    this.ready = new Promise((resolve) => resolve(this.refreshCampaignState()));
-  }
-
-  async onPageChange(callback: (state: { isNewCampaign: boolean; currentCampaign: Campaign }) => Promise<unknown>) {
-    const currentState = await this.getCurrentState();
-    await callback(currentState);
-
-    if (currentState.isNewCampaign) {
-      await this.saveCampaignToStorage(currentState.currentCampaign);
-    }
-  }
-
-  private async getCurrentState() {
-    await this.ready;
-    return {
-      isNewCampaign: this._isNewCampaign,
-      currentCampaign: this._currentCampaign,
-    };
-  }
-
-  private async refreshCampaignState() {
-    try {
-      this._currentCampaign = await this.parser.parse();
-      this._isNewCampaign = this.isNewCampaign(this._currentCampaign, await this.getCampaignFromStorage());
-    } catch {
-      // nothing to do here
-    }
-  }
-}
